+++ dataland-backend/src/main/kotlin/org/dataland/datalandbackend/services/messaging/DataOwnershipEmailMessageSender.kt
@@ -1,40 +1,47 @@
 package org.dataland.datalandbackend.services.messaging
 
 import com.fasterxml.jackson.databind.ObjectMapper
+import org.dataland.datalandbackend.repositories.StoredCompanyRepository
 import org.dataland.datalandmessagequeueutils.cloudevents.CloudEventMessageHandler
 import org.dataland.datalandmessagequeueutils.constants.ExchangeName
 import org.dataland.datalandmessagequeueutils.constants.MessageType
 import org.dataland.datalandmessagequeueutils.constants.RoutingKeyNames
 import org.dataland.datalandmessagequeueutils.messages.InternalEmailMessage
-import org.slf4j.LoggerFactory
+import org.dataland.keycloakAdapter.auth.DatalandJwtAuthentication
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Component
-import java.util.*
 
 /**
  * A class that manages generating emails messages for bulk and single data requests
  */
 @Component
-class EmailMessageSender(
+class DataOwnershipEmailMessageSender(
     @Autowired private val cloudEventMessageHandler: CloudEventMessageHandler,
     @Autowired private val objectMapper: ObjectMapper,
+    @Autowired private val companyRepository: StoredCompanyRepository
 ) {
-    private val logger = LoggerFactory.getLogger(EmailMessageSender::class.java)
-
     /**
      * Function that generates the message object for single data request mails
      */
-    fun sendSingleDataRequestInternalMessage(
-        userId: String,
+    fun sendDataOwnershipInternalEmailMessage(
+        userAuthentication: DatalandJwtAuthentication,
         datalandCompanyId: String,
-        message: InternalEmailMessage,
+        comment: String?,
+        correlationId: String,
     ) {
-        val correlationId = UUID.randomUUID().toString()
-        logger.info(
-            "User with Id $userId has submitted a single data request for company with" +
-                " Id $datalandCompanyId and correlationId $correlationId",
+        val companyName = companyRepository.findById(datalandCompanyId).get().companyName
+        val properties = mapOf(
+            "User" to userAuthentication.userDescription,
+            "Company (Dataland ID)" to datalandCompanyId,
+            "Company Name" to companyName,
+            "Comment" to comment,
         )
-
+        val message = InternalEmailMessage(
+            "Dataland Data Ownership Request",
+            "A data ownership request has been submitted",
+            "Data Ownership Request",
+            properties,
+        )
         cloudEventMessageHandler.buildCEMessageAndSendToQueue(
             objectMapper.writeValueAsString(message),
             MessageType.SendInternalEmail,
